+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### release 1.1.6 notes
+optimize attribute hover and add line empty delete funciton(enter backspace key).
+![](./examples/line-delete.gif)
+
 ### release 1.1.5 notes
 optimize close tag function、 {{}} function, add change log.